CDC PLACES geographic data on health-related burden and outcome

PLACES, a collaboration between CDC, the Robert Wood Johnson Foundation, and the CDC Foundation, allows local health departments and jurisdictions regardless of population size and urban-rural status to better understand the burden and geographic distribution of health-related outcomes in their areas and assist them in planning public health interventions.

cdc_places()

Source

https://www.cdc.gov/places/

Details

PLACES provides model-based population-level analysis and community estimates to all counties, places (incorporated and census designated places), census tracts, and ZIP Code Tabulation Areas (ZCTAs) across the United States.

PLACES is an extension of the original 500 Cities Project that provided city and census tract estimates for chronic disease risk factors, health outcomes, and clinical preventive services use for the 500 largest US cities.

They were accessed on September 25, 2021.

Download

- https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/PLACES__Local_Data_for_Better_Health__ZCTA_Data_2020_release.csv.gz

Geographic level

zipcode

See also

Examples

places <- cdc_places()

head(places)

#>    Year LocationName DataSource            Category
#> 1: 2018         1008      BRFSS Unhealthy Behaviors
#> 2: 2018         1034      BRFSS     Health Outcomes
#> 3: 2018         1035      BRFSS Unhealthy Behaviors
#> 4: 2018         1069      BRFSS     Health Outcomes
#> 5: 2018         1070      BRFSS          Prevention
#> 6: 2018         1094      BRFSS Unhealthy Behaviors
#>                                            Measure Data_Value_Unit
#> 1:            Obesity among adults aged >=18 years               %
#> 2:             Stroke among adults aged >=18 years               %
#> 3:    Current smoking among adults aged >=18 years               %
#> 4: Diagnosed diabetes among adults aged >=18 years               %
#> 5:    Mammography use among women aged 50-74 years               %
#> 6:     Binge drinking among adults aged >=18 years               %
#>     Data_Value_Type Data_Value Data_Value_Footnote_Symbol Data_Value_Footnote
#> 1: Crude prevalence       29.8                                               
#> 2: Crude prevalence        3.1                                               
#> 3: Crude prevalence       13.6                                               
#> 4: Crude prevalence        8.6                                               
#> 5: Crude prevalence       81.8                                               
#> 6: Crude prevalence       17.7                                               
#>    Low_Confidence_Limit High_Confidence_Limit TotalPopulation
#> 1:                 28.4                  31.1            1263
#> 2:                  2.7                   3.5            2021
#> 3:                 11.0                  16.2            5250
#> 4:                  8.0                   9.1            8469
#> 5:                 77.8                  85.2             629
#> 6:                 16.9                  18.4             351
#>                         Geolocation LocationID CategoryID MeasureId
#> 1: POINT (-72.95202996 42.18429555)       1008     UNHBEH   OBESITY
#> 2:   POINT (-73.014384 42.08994721)       1034    HLTHOUT    STROKE
#> 3: POINT (-72.56921312 42.35560086)       1035     UNHBEH  CSMOKING
#> 4:    POINT (-72.305378 42.1881815)       1069    HLTHOUT  DIABETES
#> 5: POINT (-72.92387449 42.51846792)       1070    PREVENT  MAMMOUSE
#> 6:  POINT (-72.1410545 42.35277567)       1094     UNHBEH     BINGE
#>    DataValueTypeID Short_Question_Text
#> 1:          CrdPrv             Obesity
#> 2:          CrdPrv              Stroke
#> 3:          CrdPrv     Current Smoking
#> 4:          CrdPrv            Diabetes
#> 5:          CrdPrv         Mammography
#> 6:          CrdPrv      Binge Drinking

# All available measures:
unique(places$Measure)

#>  [1] "Obesity among adults aged >=18 years"                                                                                                                                                          
#>  [2] "Stroke among adults aged >=18 years"                                                                                                                                                           
#>  [3] "Current smoking among adults aged >=18 years"                                                                                                                                                  
#>  [4] "Diagnosed diabetes among adults aged >=18 years"                                                                                                                                               
#>  [5] "Mammography use among women aged 50-74 years"                                                                                                                                                  
#>  [6] "Binge drinking among adults aged >=18 years"                                                                                                                                                   
#>  [7] "Arthritis among adults aged >=18 years"                                                                                                                                                        
#>  [8] "Chronic kidney disease among adults aged >=18 years"                                                                                                                                           
#>  [9] "Current asthma among adults aged >=18 years"                                                                                                                                                   
#> [10] "Cancer (excluding skin cancer) among adults aged >=18 years"                                                                                                                                   
#> [11] "Cholesterol screening among adults aged >=18 years"                                                                                                                                            
#> [12] "All teeth lost among adults aged >=65 years"                                                                                                                                                   
#> [13] "Visits to doctor for routine checkup within the past year among adults aged >=18 years"                                                                                                        
#> [14] "High blood pressure among adults aged >=18 years"                                                                                                                                              
#> [15] "Current lack of health insurance among adults aged 18-64 years"                                                                                                                                
#> [16] "Visits to dentist or dental clinic among adults aged >=18 years"                                                                                                                               
#> [17] "Physical health not good for >=14 days among adults aged >=18 years"                                                                                                                           
#> [18] "No leisure-time physical activity among adults aged >=18 years"                                                                                                                                
#> [19] "Coronary heart disease among adults aged >=18 years"                                                                                                                                           
#> [20] "Sleeping less than 7 hours among adults aged >=18 years"                                                                                                                                       
#> [21] "Cervical cancer screening among adult women aged 21-65 years"                                                                                                                                  
#> [22] "Mental health not good for >=14 days among adults aged >=18 years"                                                                                                                             
#> [23] "Chronic obstructive pulmonary disease among adults aged >=18 years"                                                                                                                            
#> [24] "High cholesterol among adults aged >=18 years who have been screened in the past 5 years"                                                                                                      
#> [25] "Taking medicine for high blood pressure control among adults aged >=18 years with high blood pressure"                                                                                         
#> [26] "Fecal occult blood test, sigmoidoscopy, or colonoscopy among adults aged 50-75 years"                                                                                                          
#> [27] "Older adult men aged >=65 years who are up to date on a core set of clinical preventive services: Flu shot past year, PPV shot ever, Colorectal cancer screening"                              
#> [28] "Older adult women aged >=65 years who are up to date on a core set of clinical preventive services: Flu shot past year, PPV shot ever, Colorectal cancer screening, and Mammogram past 2 years"