US CDC Social Vulnerability Index

The CDC's Social Vulnerability Index (SVI), created and maintained by the Geospatial Research, Analysis, and Services Program (GRASP), uses US Census data to determine the social vulnerability of every county and tract. This index ranks each county and tract based upon 15 social factors including poverty, lack of vehicle access, and crowded housing, and groups them into four related themes: 1) Socioeconomic, 2) Housing Composition and Disability, 3) Minority Status and Language, and 4) Housing and Transportation.

cdc_social_vulnerability_index()

Source

https://svi.cdc.gov/

Value

a data.frame

Details

Theme rankings: For each of the four themes, we summed the percentiles for the variables comprising each theme. We ordered the summed percentiles for each theme to determine theme-specific percentile rankings. The four summary theme ranking variables, detailed in the Data Dictionary below, are:

Socioeconomic - RPL_THEME1

Household Composition & Disability - RPL_THEME2

Minority Status & Language - RPL_THEME3

Housing Type & Transportation - RPL_THEME4

Overall tract rankings: We summed the sums for each theme, ordered the tracts, and then calculated overall percentile rankings. Please note; taking the sum of the sums for each theme is the same as summing individual variable rankings. The overall tract summary ranking variable is RPL_THEM

For detailed documentation, see https://svi.cdc.gov/Documents/Data/2018_SVI_Data/SVI2018Documentation.pdf

Download

- https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/SVI2018_US.csv.gz

Geographic level

census tract

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = cdc_social_vulnerability_index()

# column names
colnames(res)

#>   [1] "ST"         "STATE"      "ST_ABBR"    "STCNTY"     "COUNTY"    
#>   [6] "FIPS"       "LOCATION"   "AREA_SQMI"  "E_TOTPOP"   "M_TOTPOP"  
#>  [11] "E_HU"       "M_HU"       "E_HH"       "M_HH"       "E_POV"     
#>  [16] "M_POV"      "E_UNEMP"    "M_UNEMP"    "E_PCI"      "M_PCI"     
#>  [21] "E_NOHSDP"   "M_NOHSDP"   "E_AGE65"    "M_AGE65"    "E_AGE17"   
#>  [26] "M_AGE17"    "E_DISABL"   "M_DISABL"   "E_SNGPNT"   "M_SNGPNT"  
#>  [31] "E_MINRTY"   "M_MINRTY"   "E_LIMENG"   "M_LIMENG"   "E_MUNIT"   
#>  [36] "M_MUNIT"    "E_MOBILE"   "M_MOBILE"   "E_CROWD"    "M_CROWD"   
#>  [41] "E_NOVEH"    "M_NOVEH"    "E_GROUPQ"   "M_GROUPQ"   "EP_POV"    
#>  [46] "MP_POV"     "EP_UNEMP"   "MP_UNEMP"   "EP_PCI"     "MP_PCI"    
#>  [51] "EP_NOHSDP"  "MP_NOHSDP"  "EP_AGE65"   "MP_AGE65"   "EP_AGE17"  
#>  [56] "MP_AGE17"   "EP_DISABL"  "MP_DISABL"  "EP_SNGPNT"  "MP_SNGPNT" 
#>  [61] "EP_MINRTY"  "MP_MINRTY"  "EP_LIMENG"  "MP_LIMENG"  "EP_MUNIT"  
#>  [66] "MP_MUNIT"   "EP_MOBILE"  "MP_MOBILE"  "EP_CROWD"   "MP_CROWD"  
#>  [71] "EP_NOVEH"   "MP_NOVEH"   "EP_GROUPQ"  "MP_GROUPQ"  "EPL_POV"   
#>  [76] "EPL_UNEMP"  "EPL_PCI"    "EPL_NOHSDP" "SPL_THEME1" "RPL_THEME1"
#>  [81] "EPL_AGE65"  "EPL_AGE17"  "EPL_DISABL" "EPL_SNGPNT" "SPL_THEME2"
#>  [86] "RPL_THEME2" "EPL_MINRTY" "EPL_LIMENG" "SPL_THEME3" "RPL_THEME3"
#>  [91] "EPL_MUNIT"  "EPL_MOBILE" "EPL_CROWD"  "EPL_NOVEH"  "EPL_GROUPQ"
#>  [96] "SPL_THEME4" "RPL_THEME4" "SPL_THEMES" "RPL_THEMES" "F_POV"     
#> [101] "F_UNEMP"    "F_PCI"      "F_NOHSDP"   "F_THEME1"   "F_AGE65"   
#> [106] "F_AGE17"    "F_DISABL"   "F_SNGPNT"   "F_THEME2"   "F_MINRTY"  
#> [111] "F_LIMENG"   "F_THEME3"   "F_MUNIT"    "F_MOBILE"   "F_CROWD"   
#> [116] "F_NOVEH"    "F_GROUPQ"   "F_THEME4"   "F_TOTAL"    "E_UNINSUR" 
#> [121] "M_UNINSUR"  "EP_UNINSUR" "MP_UNINSUR" "E_DAYPOP"